Simple management error checks

+++ pennanalytics/snmp_flask.py
@@ -28,12 +28,18 @@ def manage():
         # The "background" task will check for the presence of the 'runtask' file.
         action = request.form['action']
         if action == 'stop':
-            os.remove(RUNFILE)
-            message = "Success" if not os.path.isfile(RUNFILE) else "Failure"
+            if not os.path.isfile(RUNFILE):
+                message = "Failure (not running already)"
+            else:
+                os.remove(RUNFILE)
+                message = "Success" if not os.path.isfile(RUNFILE) else "Failure"
 
         elif action == 'run':
-            open(RUNFILE, 'a').close()
-            message = "Success" if os.path.isfile(RUNFILE) else "Failure"
+            if os.path.isfile(RUNFILE):
+                message = "Failure (already running)"
+            else:
+                open(RUNFILE, 'a').close()
+                message = "Success" if os.path.isfile(RUNFILE) else "Failure"
 
         else:
             message = "Invalid action"
